requirements.txt
Django
djangorestframework
requests
markdown
psycopg2-binary
gunicorn
django-cors-headers==3.14.0
